add abstract PageProcessorFactoryBase class back
